Upgrade undertow version to fix CVE-2017-2666
Upgrading io.undertow:undertow-core to version 1.4.26.
Vulnerable versions: >= 1.4.0, < 1.4.17
Patched version: 1.4.17

It was discovered in Undertow that the code that parsed the HTTP request line permitted invalid characters. This could be exploited, in conjunction with a proxy that also permitted the invalid characters but with a different interpretation, to inject data into the HTTP response. By manipulating the HTTP response the attacker could poison a web-cache, perform an XSS attack, or obtain sensitive information from requests other than their own.